# docs/openapi.yaml
openapi: 3.1.0
info:
  title: API · Reservas Multi-sede (Peluquerías)
  version: "1.0"
  description: |
    API REST única que comparten la web pública, el bot de
    WhatsApp y el panel
    (doc 06). Horas en **UTC** (ISO 8601). Endpoints del panel autenticados con
    JWT (Bearer) y autorizados por rol. Errores con forma uniforme
    `{ "error": { "code", "message" } }`.
servers:
  - url: http://localhost:8000
    description: Desarrollo
...
security: []
paths:
...
  /api/v1/admin/locations:
    get:
      tags: [Panel]
      summary: Sedes
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Sedes } }
    post:
      tags: [Panel]
      summary: Crear sede (admin_cadena)
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [name, slug], properties: { name: { type: string }, slug: { type: string }, address: { type: [string, "null"] }, phone: { type: [string, "null"] }, timezone: { type: string } } } } } }
      responses: { "201": { description: Creada }, "409": { $ref: "#/components/responses/Error" } }
  /api/v1/admin/locations/{id}:
    patch:
      tags: [Panel]
      summary: Editar sede (admin_cadena)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { "200": { description: Actualizada } }
  /api/v1/admin/locations/{id}/branding:
    get:
      tags: [Panel]
      summary: Branding de la sede
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { "200": { description: Branding } }
    patch:
      tags: [Panel]
      summary: Editar branding (custom_domain solo admin_cadena)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { logo_url: { type: [string, "null"] }, color_primary: { type: [string, "null"] }, color_accent: { type: [string, "null"] }, font_family: { type: [string, "null"] }, custom_domain: { type: [string, "null"] }, extra: { type: [object, "null"] } } } } } }
      responses: { "200": { description: Actualizado } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Error con forma uniforme
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: SLOT_TAKEN }
            message: { type: string }
...